+++ src/components/organisms/ValidationSettings/CustomValidationPaneHeading/CustomValidationPaneHeading.styled.tsx
