+++ src/components/organisms/ValidationSettings/CustomValidationPaneHeading/index.ts
@@ -1,0 +1,1 @@
+export {default} from './CustomValidationPaneHeading';
