--- src/components/organisms/ValidationSettings/ValidationOpenPolicyAgent/ValidationOpenPolicyAgent.styled.tsx
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-
-export const Description = styled.p`
-  margin-bottom: 0;
-`;
-
-export const DescriptionContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 5px;
-  padding: 16px;
-`;
-
-export const DescriptionActions = styled.div`
-  display: flex;
-`;
