--- src/components/organisms/ValidationSettings/ValidationOpenPolicyAgent/index.ts
@@ -1,1 +1,0 @@
-export {default} from './ValidationOpenPolicyAgent';
