--- src/components/organisms/ValidationSettings/ValidationPaneHeading/index.ts
@@ -1,1 +1,0 @@
-export {default} from './ValidationPaneHeading';
